[Tests] Reorganize tests files & tests
 - Rename tests files for a better isolation of the category of tests
 - use pytest-order to order tests relatively to each other (currently
   incomplete due to lacking behavior from pytest-order)
 - Update circleCI workflow to reflect the previous changes
