+++ requirements.txt
@@ -1,2 +1,3 @@
 selenium
 pytest
+pytest-order
