--- tests/func/test_010_utilities.py
@@ -1,6 +1,0 @@
-import pytest
-
-def test_010_utilities(controller, reader_driver):
-    reader_driver.load_random()
-    reader_driver.prev_page()
-    reader_driver.next_page()
