+++ tests/func/test_load.py
@@ -4,13 +4,13 @@ from .utils.extension import Extension
 
 EXT = Extension()
 
-TEST_URLS = EXT.supported_readers + [
+TEST_URLS = EXT.supported_readers[:1] + [
     pytest.param('https://www.qwant.com/', marks=pytest.mark.xfail)
 ]
 
 
 @pytest.mark.parametrize('reader_url', TEST_URLS)
-def test_001_load(controller, reader_url):
+def test_webext_loads(controller, reader_url):
     # Load the reader's URL
     controller.driver.get(reader_url)
     # Ensure that the frame with ID 'BmcSidePanel' is loaded.
